requirements.txt
Django
neomodel
psycopg2
djangorestframework